+++ CHANGELOG.rst
@@ -2,11 +2,12 @@
 Changelog for package grepros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-0.3.0 (2021-11-19)
+0.3.0 (2021-11-21)
 -------------------
 * add embag support, for faster grepping
 * match anything by default if no patterns given
 * fix breaking too early on --max-per-topic
+* fix adding topic views to existing SQLite output database
 
 0.2.4 (2021-11-18)
 -------------------